requirements.txt
alembic==1.10.2
anyio==3.6.2
asyncpg==0.27.0
attrs==23.1.0
cachetools==5.3.1
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.3
click-plugins==1.1.1
cligj==0.7.2
colorama==0.4.6
contourpy==1.1.0
cycler==0.11.0
dnspython==2.3.0
email-validator==2.0.0.post2
et-xmlfile==1.1.0
fastapi==0.95.1
fastapi-async-sqlalchemy==0.3.12
fastapi-pagination==0.12.2
Fiona==1.9.3
fonttools==4.42.0
GeoAlchemy2==0.13.2
geojson==3.0.1
geopandas==0.13.0
google-api-core==2.11.1
google-auth==2.16.0
google-cloud==0.34.0
google-cloud-core==2.3.3
google-cloud-storage==2.10.0
google-cloud-tasks==2.13.0
google-crc32c==1.5.0
google-resumable-media==2.5.0
googleapis-common-protos==1.59.1
greenlet==2.0.2
grpc-google-iam-v1==0.12.6
grpcio==1.56.0
grpcio-status==1.56.0
gunicorn==20.1.0
h11==0.14.0
idna==3.4
Jinja2==3.1.2
kiwisolver==1.4.4
Mako==1.2.4
MarkupSafe==2.1.2
matplotlib==3.7.2
munch==2.5.0
mypy==1.5.1
mypy-extensions==1.0.0
Naked==0.1.32
numpy==1.24.3
openpyxl==3.1.2
packaging==23.1
pandas==2.0.1
pdfkit==1.0.0
Pillow==10.0.0
proto-plus==1.22.3
protobuf==4.23.4
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycryptodome==3.19.0
pydantic==1.10.7
pygeos==0.14
pyparsing==3.0.9
pyproj==3.5.0
python-dateutil==2.8.2
python-dotenv==1.0.0
python-multipart==0.0.6
pytz==2023.3
PyYAML==6.0.1
requests==2.30.0
roman==4.1
rsa==4.9
shapely==2.0.1
shellescape==3.8.1
six==1.16.0
sniffio==1.3.0
SQLAlchemy==1.4.41
sqlalchemy2-stubs==0.0.2a34
sqlmodel==0.0.8
starlette==0.26.1
stringcase==1.2.0
tomli==2.0.1
typer==0.9.0
typing_extensions==4.5.0
tzdata==2023.3
urllib3==1.26.16
uvicorn==0.22.0
